Tests/reorganize: move setFrom tests to own file
